requirements.txt
pyTelegramBotAPI==4.3.1
requests==2.28.1
pandas==1.5.3
SQLAlchemy==1.4.36
boto3==1.24.12
pymongo==4.7.2
deep-translator>=1.10,<2.0
psycopg2-binary>=2.9,<3.0
